Create modules to create a project
