+++ js/index.js
@@ -2,8 +2,10 @@
 /* eslint-disable no-plusplus */
 /* eslint-disable no-unused-vars */
 import createPopupProject from './modules/createPopupProject.js';
+import createProject from './modules/createProject.js';
 import { formListener, displayData } from './modules/form.js';
 import projectObj from './modules/projectsObj.js';
+import { enableScroll, disableScroll } from './modules/enableDisableScroll.js';
 
 window.sr = ScrollReveal();
 
@@ -35,52 +37,7 @@ sr.reveal(form, {
   delay: 200,
 });
 
-const projectData = projectObj();
-
-function createProject(index) {
-  const { name, mobileSnapshot, technologies } = projectData[index];
-  const projectSample = document.createElement('div');
-  const projectDetails = document.createElement('div');
-  const projectName = document.createElement('h3');
-  const projectUl = document.createElement('ul');
-  const seeProjectBtn = document.createElement('button');
-
-  projectSample.classList.add('project-sample');
-  projectDetails.classList.add('details');
-  projectUl.classList.add('tech-used');
-  seeProjectBtn.classList.add('green-btn', 'popup-open-btn');
-
-  projectName.textContent = name;
-  seeProjectBtn.textContent = 'See Project';
-
-  projectSample.style.backgroundImage = `url('../${mobileSnapshot}')`;
-  projectSample.style.backgroundPosition = 'center';
-  projectSample.style.backgroundSize = 'cover';
-
-  seeProjectBtn.type = 'button';
-  seeProjectBtn.id = `project-${index}`;
-
-  for (let i = 0; i < technologies.length; i++) {
-    const projectLi = document.createElement('li');
-    const a = document.createElement('a');
-
-    projectLi.classList.add('tech-used-item');
-
-    a.setAttribute('href', '#');
-    a.textContent = technologies[i];
-    projectLi.append(a);
-    projectUl.append(projectLi);
-  }
-
-  projectDetails.append(projectName, projectUl, seeProjectBtn);
-  projectSample.append(projectDetails);
-  projectsContainer.append(projectSample);
-  return projectsContainer;
-}
-
-const enableScroll = () => {
-  window.onscroll = () => {};
-};
+const projectData = projectObj;
 
 const displayCourse = () => {
   projectData.forEach((_, index) => {
@@ -89,15 +46,6 @@ const displayCourse = () => {
 };
 
 displayCourse();
-
-const disableScroll = () => {
-  const scrollTop = window.pageYOffset || document.documentElement.scrollTop;
-  const scrollLeft = window.pageXOffset || document.documentElement.scrollLeft;
-
-  window.onscroll = () => {
-    window.scrollTo(scrollLeft, scrollTop);
-  };
-};
 
 const openPopup = (id) => {
   const pop = createPopupProject(id);
@@ -126,10 +74,6 @@ document.addEventListener('click', (event) => {
     openPopup(index);
   }
 });
-
-const openMenu = () => {
-  mobileMenu.style.display = 'block';
-};
 
 const closeMenu = () => {
   mobileMenu.style.display = 'none';
