+++ js/modules/enableDisableScroll.js
@@ -1,0 +1,14 @@
+const enableScroll = () => {
+  window.onscroll = () => {};
+};
+
+const disableScroll = () => {
+  const scrollTop = window.pageYOffset || document.documentElement.scrollTop;
+  const scrollLeft = window.pageXOffset || document.documentElement.scrollLeft;
+
+  window.onscroll = () => {
+    window.scrollTo(scrollLeft, scrollTop);
+  };
+};
+
+export { enableScroll, disableScroll };
